feat: add .useVuetifyColors
